Date: Wed, 28 Jan 2004 02:37:04 +0100 From: Max Horn Subject: Small libtool patch (framework related)
here is a small patch for ltmain.sh. It corrects a problems with the
libSDL.la file, which prevented applications using a version of libtool
w/o framework support (read: all currently released libtool version) to
not being able to link with SDL anymore. I also notified the libtool
maintainers, as the same problem exists in libtool CVS.