Eric Wing to Sam
Included is the fix for the missing SDL_revision.h in the framework.

Also, I include the project template fixes for the prefix header files. But since this is in SVN directly, you must rename the files named
SDLApp_Prefix.pch
to
___PROJECTNAMEASIDENTIFIER____Prefix.pch

in the TemplatesForXcodeSnowLeopard and TemplatesForXcodeLeopard directories. (Do not rename the Tiger ones.)